+++ src/components/basic_table/basic_table.js
@@ -384,7 +384,7 @@ export class EuiBasicTable extends Component {
     );
 
     const table = this.renderTable();
-    const paginationBar = this.renderPaginationBar(items);
+    const paginationBar = this.renderPaginationBar(items.length);
 
     return (
       <div className={classes} {...rest}>
@@ -1049,9 +1049,9 @@ export class EuiBasicTable extends Component {
     return profile.align;
   }
 
-  renderPaginationBar(items) {
+  renderPaginationBar(itemsLength) {
     const { error, pagination, onChange } = this.props;
-    if (!error && pagination && items.length > 0) {
+    if (!error && pagination && itemsLength > 0) {
       if (!onChange) {
         throw new Error(`The Basic Table is configured with pagination but [onChange] is
         not configured. This callback must be implemented to handle pagination changes`);